Home work 4

See merge request spinoza0/tfs-android-spring-2023!8

+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/data/model/ReactionParamDto.kt
@@ -1,5 +1,5 @@
 package com.spinoza.messenger_tfs.data.model
 
 data class ReactionParamDto(
-    val usersIds: List<Int>,
+    val usersIds: List<Long>,
 )
